+++ src/main/java/com/pekar/angelblock/blocks/tile_entities/monsters/EnderDragon.java
@@ -10,7 +10,7 @@ import net.minecraft.world.item.Items;
 
 public class EnderDragon extends Monster
 {
-    public EnderDragon(byte id)
+    public EnderDragon(int id)
     {
         super(id);
     }
