+++ src/main/java/com/pekar/angelblock/blocks/tile_entities/monsters/IMonster.java
@@ -8,7 +8,7 @@ import net.minecraft.world.item.Item;
 
 public interface IMonster
 {
-    byte getId();
+    int getId();
     boolean belongs(LivingEntity livingEntity);
     Item getActionItem();
     EntityType<? extends Entity> getEntityType();
