+++ src/main/java/com/pekar/angelblock/blocks/tile_entities/monsters/Zombie.java
@@ -10,7 +10,7 @@ import net.minecraft.world.item.Items;
 
 public class Zombie extends Monster
 {
-    public Zombie(byte id)
+    public Zombie(int id)
     {
         super(id);
     }
